change location of buildDynamicURL
